separate setupperandlowerelements from staffgroup element
