+++ src/write/abc_staff_group_element.js
@@ -106,91 +106,6 @@ StaffGroupElement.prototype.setStaffLimits = function (voice) {
 	this.setLimit('dynamicHeightBelow', voice);
 };
 
-StaffGroupElement.prototype.setUpperAndLowerElements = function(renderer) {
-	// Each staff already has the top and bottom set, now we see if there are elements that are always on top and bottom, and resolve their pitch.
-	// Also, get the overall height of all the staves in this group.
-	var lastStaffBottom;
-	for (var i = 0; i < this.staffs.length; i++) {
-		var staff = this.staffs[i];
-		// the vertical order of elements that are above is: tempo, part, volume/dynamic, ending/chord, lyric
-		// the vertical order of elements that are below is: lyric, chord, volume/dynamic
-		var positionY = {
-			tempoHeightAbove: 0,
-			partHeightAbove: 0,
-			volumeHeightAbove: 0,
-			dynamicHeightAbove: 0,
-			endingHeightAbove: 0,
-			chordHeightAbove: 0,
-			lyricHeightAbove: 0,
-
-			lyricHeightBelow: 0,
-			chordHeightBelow: 0,
-			volumeHeightBelow: 0,
-			dynamicHeightBelow: 0
-		};
-
-		if (/*ABCJS.write.debugPlacement*/false) {
-			staff.originalTop = staff.top; // This is just being stored for debugging purposes.
-			staff.originalBottom = staff.bottom; // This is just being stored for debugging purposes.
-		}
-
-		if (staff.specialY.lyricHeightAbove) { staff.top += staff.specialY.lyricHeightAbove; positionY.lyricHeightAbove = staff.top; }
-		if (staff.specialY.chordHeightAbove) { staff.top += staff.specialY.chordHeightAbove; positionY.chordHeightAbove = staff.top; }
-		if (staff.specialY.endingHeightAbove) {
-			if (staff.specialY.chordHeightAbove)
-				staff.top += 2;
-			else
-				staff.top += staff.specialY.endingHeightAbove;
-			positionY.endingHeightAbove = staff.top;
-		}
-		if (staff.specialY.dynamicHeightAbove && staff.specialY.volumeHeightAbove) {
-			staff.top += Math.max(staff.specialY.dynamicHeightAbove, staff.specialY.volumeHeightAbove);
-			positionY.dynamicHeightAbove = staff.top;
-			positionY.volumeHeightAbove = staff.top;
-		} else if (staff.specialY.dynamicHeightAbove) {
-			staff.top += staff.specialY.dynamicHeightAbove; positionY.dynamicHeightAbove = staff.top;
-		} else if (staff.specialY.volumeHeightAbove) { staff.top += staff.specialY.volumeHeightAbove; positionY.volumeHeightAbove = staff.top; }
-		if (staff.specialY.partHeightAbove) { staff.top += staff.specialY.partHeightAbove; positionY.partHeightAbove = staff.top; }
-		if (staff.specialY.tempoHeightAbove) { staff.top += staff.specialY.tempoHeightAbove; positionY.tempoHeightAbove = staff.top; }
-
-		if (staff.specialY.lyricHeightBelow) { positionY.lyricHeightBelow = staff.bottom; staff.bottom -= staff.specialY.lyricHeightBelow; }
-		if (staff.specialY.chordHeightBelow) { positionY.chordHeightBelow = staff.bottom; staff.bottom -= staff.specialY.chordHeightBelow; }
-		if (staff.specialY.volumeHeightBelow && staff.specialY.dynamicHeightBelow) {
-			positionY.volumeHeightBelow = staff.bottom;
-			positionY.dynamicHeightBelow = staff.bottom;
-			staff.bottom -= Math.max(staff.specialY.volumeHeightBelow, staff.specialY.dynamicHeightBelow);
-		} else if (staff.specialY.volumeHeightBelow) {
-			positionY.volumeHeightBelow = staff.bottom; staff.bottom -= staff.specialY.volumeHeightBelow;
-		} else if (staff.specialY.dynamicHeightBelow) {
-			positionY.dynamicHeightBelow = staff.bottom; staff.bottom -= staff.specialY.dynamicHeightBelow;
-		}
-
-		if (/*ABCJS.write.debugPlacement*/false)
-			staff.positionY = positionY; // This is just being stored for debugging purposes.
-
-		for (var j = 0; j < staff.voices.length; j++) {
-			var voice = this.voices[staff.voices[j]];
-			voice.setUpperAndLowerElements(positionY);
-		}
-		// We might need a little space in between staves if the staves haven't been pushed far enough apart by notes or extra vertical stuff.
-		// Only try to put in extra space if this isn't the top staff.
-		if (lastStaffBottom !== undefined) {
-			var thisStaffTop = staff.top - 10;
-			var forcedSpacingBetween = lastStaffBottom + thisStaffTop;
-			var minSpacingInPitches = renderer.spacing.systemStaffSeparation/spacing.STEP;
-			var addedSpace = minSpacingInPitches - forcedSpacingBetween;
-			if (addedSpace > 0)
-				staff.top += addedSpace;
-		}
-		lastStaffBottom = 2 - staff.bottom; // the staff starts at position 2 and the bottom variable is negative. Therefore to find out how large the bottom is, we reverse the sign of the bottom, and add the 2 in.
-
-		// Now we need a little margin on the top, so we'll just throw that in.
-		//staff.top += 4;
-		//console.log("Staff Y: ",i,heightInPitches,staff.top,staff.bottom);
-	}
-	//console.log("Staff Height: ",heightInPitches,this.height);
-};
-
 StaffGroupElement.prototype.finished = function() {
 	for (var i=0;i<this.voices.length;i++) {
 		if (!this.voices[i].layoutEnded()) return false;
